requirements.txt
nats-py
PyYAML

# requests
# fastapi[all]
# pydantic
# uvicorn
# sqlalchemy
# RPi.GPIO
